+++ src/ts/toolbar/Fullscreen.ts
@@ -23,6 +23,7 @@ export class Fullscreen extends MenuItem {
                         svgElement.className = svgElement.className.replace("__s", "__n");
                     }
                 });
+                vditor.counter.element.className = vditor.counter.element.className.replace("__s", "__n");
             } else {
                 this.innerHTML = menuItem.icon || contractSVG;
                 vditor.element.classList.add("vditor--fullscreen");
@@ -32,6 +33,7 @@ export class Fullscreen extends MenuItem {
                         svgElement.className = svgElement.className.replace("__n", "__s");
                     }
                 });
+                vditor.counter.element.className = vditor.counter.element.className.replace("__n", "__s");
             }
 
             if (vditor.devtools) {